requirements.txt
blinker==1.9.0
boto3==1.39.3
botocore==1.39.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.3.1
pandas==2.3.0
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
requests==2.32.4
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
spotipy==2.25.1
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
